+++ courses/questionnaire/src/main/java/by/academy/questionnaire/database/UserDAO.kt
@@ -12,7 +12,7 @@ import by.academy.questionnaire.database.entity.UserEntity
 interface UserDAO {
 
     @Query("SELECT count(1) FROM user")
-    fun size(): Long
+    fun size(): Int
 
     @Query("SELECT * FROM user order by u_id")
     fun getAll(): List<UserEntity>
